[#57] Have to connect API
API 연결 Based
