+++ src/components/SelectCountry/index.tsx
@@ -1,120 +1,119 @@
-import React, { useState } from "react";
-import CountryDropDown from "./CountryDropDown";
+import axios from "axios";
+import { useRouter } from "next/router";
+import { ChangeEvent, FormEvent, useState } from "react";
+import { API_BASED_URL } from "src/constants/apiUrl";
+// import { ICountryHelp } from "src/pages/DontWorry/Help";
 import styled from "styled-components";
 
-function SelectCountry() {
-  const [showDropDown, setShowDropDown] = useState<boolean>(false);
-  const [selectCountry, setSelectCountry] = useState<string>("");
-  const countries = () => {
-    return ["KR, Korea", "US, United States", "CN, China", "JP, Japan"];
+interface ICountryHelp {
+  country_name: string;
+  number: string | number;
+  summary: string;
+}
+
+function SelectCountryT() {
+  const router = useRouter();
+  const [selectedCountry, setSelectedCountry] = useState<String>("");
+  const [country, setCountry] = useState<String>("");
+  const [helpInform, setHelpInform] = useState<ICountryHelp[]>([]);
+
+  const selectChange = (e: ChangeEvent<HTMLSelectElement>) => {
+    const value = e.target.value;
+    setSelectedCountry(value);
+    setCountry(value);
   };
 
-  const toggleDropDown = () => {
-    setShowDropDown(!showDropDown);
-  };
+  const submitCountry = async (e: FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    {
+      await axios
+        .get(`${API_BASED_URL}/help/${country}`, {
+          params: {
+            name: "country",
+          },
+          withCredentials: true,
+        })
+        .then((res) => {
+          console.log(res);
+          console.log(res.data);
+          // 각 나라 헬프 정보 출력 확인.
+          console.log(res.data[0]);
 
-  const outFoucsHandler = (e: React.FocusEvent<HTMLButtonElement>) => {
-    if (e.currentTarget === e.target) {
-      setShowDropDown(false);
+          // 배열타입으로 인터페이스에 res.data 정보 담고 props 형태로 보내줄려는 방법?
+          // 이게 되나?
+          setHelpInform(res.data);
+          console.log(setHelpInform);
+
+          // router.push("/DontWorry/Help");
+          // 스토리지에 저장해서 갖고오는 방법? 객체가 [Object object]로 표시됨.
+          // sessionStorage.setItem("countryHelp", res.data);
+        })
+        .catch((err) => {
+          console.log(err);
+          alert("잘못된 페이지");
+        });
     }
   };
 
-  const countrySelection = (country: string) => {
-    setSelectCountry(country);
-  };
-
   return (
-    <>
-      <div>
-        <div>
-          {selectCountry
-            ? `You selected ${selectCountry}`
-            : "Select Your Country"}
-        </div>
-      </div>
-      <StyledButtonDiv>
-        <StyledButton
-          className={showDropDown ? "active" : undefined}
-          onClick={(): void => toggleDropDown()}
-          onBlur={(e: React.FocusEvent<HTMLButtonElement>): void =>
-            outFoucsHandler(e)
-          }
-        >
-          <div>
-            {/* {selectCountry ? "Selected: " + selectCountry : "Select ..."} */}
-            {selectCountry ? "" + selectCountry : "Select ..."}
-          </div>
-          {showDropDown && (
-            <CountryDropDown
-              countries={countries()}
-              showDropDown={false}
-              toggleDropDown={() => toggleDropDown()}
-              countrySelection={countrySelection}
-            />
-          )}
-        </StyledButton>
-      </StyledButtonDiv>
-    </>
+    <StyledSelectDiv>
+      <StyledSelect onChange={selectChange} defaultValue={"country" || ""}>
+        <option value="country" disabled>
+          Choose Country ...
+        </option>
+        <option value="USA">🇺🇸 | United States</option>
+        <option value="KOREA">🇰🇷 | Korea</option>
+        <option value="JAPAN">🇯🇵 | Japan</option>
+        <option value="CHINA">🇨🇳 | China</option>
+      </StyledSelect>
+      {selectedCountry && <h2>{selectedCountry}</h2>}
+      {/* <form onSubmit={submitCountry}>
+        <button type="submit">선택 나라 헬프 페이지 이동</button>
+      </form> */}
+    </StyledSelectDiv>
   );
 }
 
-export default SelectCountry;
+export default SelectCountryT;
 
-const StyledButton = styled.button`
-  position: relative;
-  padding: 0.375rem 0.88rem;
-  white-space: nowrap;
-  vertical-align: middle;
-  user-select: none;
-  -webkit-user-select: none;
-
-  color: ${({ theme }) => theme.color.white};
-  background-color: ${({ theme }) => theme.color.grey};
-
-  font-size: 0.85rem;
-  text-align: left;
-  line-height: 1.5;
-  font-weight: ${({ theme }) => theme.fontWeight.normal};
-
-  border: 0.125rem solid transparent;
-  border-radius: ${({ theme }) => theme.borderRadius.imgCard};
-  cursor: pointer;
-  :hover {
-    color: ${({ theme }) => theme.color.white};
-    background-color: ${({ theme }) => theme.color.blueGreen};
-  }
-  .active {
-    color: ${({ theme }) => theme.color.white};
-    background-color: ${({ theme }) => theme.color.blueGreen};
-  }
-  .dropdown {
-    position: absolute;
-    top: 2.375rem;
-    left: -0.0625rem;
-    border: 0.0625rem solid rgb(197, 197, 197);
-    background: ${({ theme }) => theme.color.white};
-    padding: 0;
-    color: ${({ theme }) => theme.color.blueGreen};
-    text-align: left;
-    border-radius: 0.25rem;
-  }
-  .dropdown > p {
-    margin: 0;
-    padding: 0.375rem 0.88rem;
-    border-bottom: 0.0625rem solid ${({ theme }) => theme.color.white};
-    min-width: 5rem;
-  }
-  .dropdown > p:hover {
-    color: ${({ theme }) => theme.color.white};
-    background-color: ${({ theme }) => theme.color.blueGreen};
-  }
-  .dropdown > p:last-child {
-    border-bottom: 0 none;
-  }
+const StyledSelectDiv = styled.div`
+  margin-top: 4rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 `;
 
-const StyledButtonDiv = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
+const StyledSelect = styled.select`
+  width: 10rem;
+  padding: 0.5rem;
+
+  border: 0.0425rem solid ${({ theme }) => theme.color.grey100};
+  border-radius: ${({ theme }) => theme.borderRadius.imgCard};
+
+  background-color: ${({ theme }) => theme.color.background};
+  color: ${({ theme }) => theme.color.white};
+
+  cursor: pointer;
+
+  :hover {
+    border-color: ${({ theme }) => theme.color.blueGreen};
+    transition: all 0.5s;
+  }
+
+  :focus {
+    border-color: ${({ theme }) => theme.color.grey100};
+    box-shadow: 0 0 1px 1px rgba(151, 200, 252, 0.4);
+    color: ${({ theme }) => theme.color.white};
+    outline: none;
+  }
+
+  /* > option[value="US"] {
+    background-image: url("/images/america.png");
+    background-repeat: no-repeat;
+  } */
+
+  option:checked {
+    background-color: ${({ theme }) => theme.color.grey100};
+    color: ${({ theme }) => theme.color.white};
+  }
 `;
